Docker build for backend

+++ database/CrawlLinks.sql
@@ -1,8 +1,8 @@
 START TRANSACTION;
 
 CREATE TABLE CrawlLinks (
-  parent_id int NOT NULL,
-  son_id int NOT NULL
+  parent_id bigint NOT NULL,
+  son_id bigint NOT NULL
 );
 
 COMMIT;
